Fix implementation of Wigner thermal conductivity: - employ the smooth (Wallace) phase, which ensures size consistency (see https://link.aps.org/doi/10.1103/PhysRevX.12.041011 , in particular Sec VI); - consider the Wigner velocity operator also at q=0, a condition that is crucial to have a size-consistent formulation and describe the disordered-solid limit (see https://www.nature.com/articles/s41524-023-01033-4 ) .
The current implementation is limited to the RTA level and outputs only the total Wigner conductivity, without decomposing it into population and coherences contributions. The decomposition between populations and coherences will be implemented in the future, along with the possibility to output the Wigner conductivity also with the full exact solution of the BTE. The code gives results compatible with phono3py.